requirements.txt
pyinstaller~=6.9.0
openpyxl~=3.1.5
pandas~=2.2.3
python_socks~=2.5.3
telethon~=1.36.0
ttkthemes~=3.2.2
